+++ instructors/save_course.php
@@ -2,11 +2,12 @@
 //include db_connect function
 include "./../includes/db_connect.php";
 include "./../includes/functions.php";
+include "./../includes/server-funcs.php";
+include "./../includes/views.php";
 
 session_start();
 if(isset($_SESSION["owner_id"])){
 $owner_id = $_SESSION["owner_id"];
-$lecturer_db = $_SESSION["lecturer_db"];
 
 $heading = ""; $display = "";
 if(isset($_GET["save_courses"]) || isset($_POST["edit_course"])){
@@ -20,11 +21,11 @@ if(isset($_GET["save_courses"]) || isset($_POST["edit_course"])){
 			$course_title = ""; $course_description = ""; $course_unit; $course_unit = ""; $course_code = ""; $action = ""; $view_courses = "";
 		}	else 	{
 			$course_id = $_POST["course_id"][0];
-			$course_detail = get_course_code($course_id, 3, $lecturer_db);
+			$course_detail = get_course_code($course_id, 3);
 			if(empty($course_detail)){
 				$diplay = "<p>The course you want to edit does not exist</p>";
 			}	else	{
-				$course_code = get_course_code($course_id, 1, $lecturer_db);
+				$course_code = get_course_code($course_id, 1);
 				$heading = "<h1>Your are editing $course_code</h1>";
 				$course_code = "<input type = \"hidden\" name = \"course_code\" value = \"$course_id\" />";
 				$course_title .= "<input type = \"text\" name = \"course_title\" value = \"$course_detail[1]\" />";
@@ -71,24 +72,24 @@ if(isset($_POST["save_course"]) || isset($_POST["update_course"])){
 		$course_description = mysqli_real_escape_string($mysqli, trim($_POST["course_description"]));
 		$course_unit = mysqli_real_escape_string($mysqli, trim($_POST["course_unit"])); 
 		if(isset($_POST["update_course"])){
-			$query_string = "update courses set course_title = \"$course_title\", course_description = \"$course_description\", unit = \"$course_unit\" where course_id = \"$course_code\""; 	//$course_code is actually the course_id
-			$course_code = get_course_code($course_code, 1, $lecturer_db);
+			$query_string = "update courses set course_title = \"$course_title\", course_description = \"$course_description\", unit = \"$course_unit\" where course_id = \"$course_code\" and lec_id = \"$owner_id\""; 	//$course_code is actually the course_id
+			$course_code = get_course_code($course_code, 1);
 			$success_string = "<p>You have successfully updated $course_code</p>";
 			$failure_string = "<p>The course could not be added </p>";
 		}	
 		if(isset($_POST["save_course"])){
-			$query_string = "select * from courses where course_code = \"$course_code\"";
-			run_query($query_string, $lecturer_db);
+			$query_string = "select * from courses where course_code = \"$course_code\" and lec_id = \"$owner_id\"";
+			run_query($query_string);
 			if($row_num2 == 1){
 				$display = "<p>This Course has already been save</p><br/>";
 				$display .="<a href = \"$_SERVER[PHP_SELF]?view_courses=yes\" id = \"viewCourses\" class = \"btn btn-primary\" >View courses</a>";
 			}	else	{
-				$query_string = "insert into courses values (null, \"$course_code\", \"$course_title\", \"$course_description\", \"$course_unit\")";
+				$query_string = "insert into courses values (null, \"$owner_id\", \"$course_code\", \"$course_title\", \"$course_description\", \"$course_unit\")";
 				$success_string = "<p>$course_code has been saved successfully</p>";
 				$failure_string = "<p>The course could not be added </p>";
 			}
 		}
-		run_query($query_string, $lecturer_db);
+		run_query($query_string);
 		if($row_num2 == 1){
 			$display = $success_string;
 		}	else {
@@ -100,9 +101,9 @@ if(isset($_POST["save_course"]) || isset($_POST["update_course"])){
 
 
 if(isset($_GET["view_courses"])){
-	$query_string = "select * from courses";
+	$query_string = "select * from courses where lec_id = \"$owner_id\"";
 	$fields = array ("course_id", "Course code", "Course Title", "Course description", "unit");
-	$values = get_course_code("course_code", 1, $lecturer_db, $query_string);	//get the course details
+	$values = get_course_code("course_code", 1, $query_string);	//get the course details
 	if(empty($values)){
 		$display = "<p>You have not save any course yet<br />go to the <a href = \"$_SERVER[PHP_SELF]?save_courses=yes\">
 				Save Courses</a> page and begin adding courses</p>";
@@ -112,8 +113,8 @@ if(isset($_GET["view_courses"])){
 		$heading = "<h1>Courses you are taking are </h1>";
 		$display  = "<form method = \"POST\" action = \"$_SERVER[PHP_SELF]\" name = \"available_courses\" >";
 		$display .= $table_values;
-		$display .= "<input type = \"submit\" class = \"btn btn-success\" id = \"deleteCourse\" name = \"delete_course\" value = \"DELETE\" />";
-		$display .= "<input type = \"submit\" class = \"btn btn-danger\" id = \"editCourse\" name = \"edit_course\" value = \"EDIT\" />";
+		$display .= "<input type = \"submit\" class = \"btn btn-success\" id = \"editCourse\" name = \"edit_course\" value = \"EDIT\" />";
+		$display .= "<input type = \"submit\" class = \"btn btn-danger\" id = \"deleteCourse\" name = \"delete_course\" value = \"DELETE\" />";
 		$display .= "</form>";
 		$display .= "<p>You want to <a href = \"$_SERVER[PHP_SELF]?save_courses=yes\" id = \"saveCourse\" class = \"btn btn-primary\" >add a course</a></p>";
 	}
@@ -127,8 +128,8 @@ if(isset($_POST["delete_course"])){
 		$course_id = $_POST["course_id"][0];
 		$display = "want to delete a course";
 		//admin_connect();
-		$query_string = "delete from courses where course_id = \"$course_id\"";
-		run_query($query_string, $lecturer_db);
+		$query_string = "delete from courses where course_id = \"$course_id\" and lec_id = \"$owner_id\"";
+		run_query($query_string);
 		if($row_num2 == 1){
 			$display = "<p>The selected course has been deleted from the record</p>";
 		}	else 	{
